+++ internal/table/retry.go
@@ -18,7 +18,7 @@ type sessionPool interface {
 	closer.Closer
 
 	Stats() pool.Stats
-	With(ctx context.Context, f func(ctx context.Context, s *session) error, opts ...retry.Option) error
+	With(ctx context.Context, f func(ctx context.Context, s *session) error, preferredNodeId uint32, opts ...retry.Option) error
 }
 
 func do(
@@ -27,6 +27,7 @@ func do(
 	config *config.Config,
 	op table.Operation,
 	onAttempt func(err error),
+	preferredNodeId uint32,
 	opts ...retry.Option,
 ) (err error) {
 	return retryBackoff(ctx, pool,
@@ -54,6 +55,7 @@ func do(
 
 			return nil
 		},
+		preferredNodeId,
 		opts...,
 	)
 }
@@ -62,6 +64,7 @@ func retryBackoff(
 	ctx context.Context,
 	pool sessionPool,
 	op table.Operation,
+	preferredNodeId uint32,
 	opts ...retry.Option,
 ) error {
 	return pool.With(ctx, func(ctx context.Context, s *session) error {
@@ -72,7 +75,7 @@ func retryBackoff(
 		}
 
 		return nil
-	}, opts...)
+	}, preferredNodeId, opts...)
 }
 
 func (c *Client) retryOptions(opts ...table.Option) *table.Options {