+++ retry/retry.go
@@ -19,14 +19,15 @@ import (
 type retryOperation func(context.Context) (err error)
 
 type retryOptions struct {
-	label       string
-	call        call
-	trace       *trace.Retry
-	idempotent  bool
-	stackTrace  bool
-	fastBackoff backoff.Backoff
-	slowBackoff backoff.Backoff
-	budget      budget.Budget
+	label           string
+	call            call
+	trace           *trace.Retry
+	idempotent      bool
+	stackTrace      bool
+	fastBackoff     backoff.Backoff
+	slowBackoff     backoff.Backoff
+	budget          budget.Budget
+	preferredNodeId uint32
 
 	panicCallback func(e interface{})
 }